+++ pallas-applying/src/utils/environment.rs
@@ -2,10 +2,11 @@
 //! era.
 use pallas_primitives::{
     alonzo::{
-        Coin, CostMdls, ExUnitPrices, ExUnits, Nonce, ProtocolVersion, RationalNumber, UnitInterval,
+        Coin, CostModels, ExUnitPrices, ExUnits, Nonce, ProtocolVersion, RationalNumber,
+        UnitInterval,
     },
-    babbage::CostMdls as BabbageCostMdls,
-    conway::{CostMdls as ConwayCostMdls, Epoch},
+    babbage::CostModels as BabbageCostModels,
+    conway::{CostModels as ConwayCostModels, Epoch},
 };
 
 #[allow(clippy::large_enum_variant)]
@@ -100,7 +101,7 @@ pub struct AlonzoProtParams {
     pub protocol_version: ProtocolVersion,
     pub min_pool_cost: Coin,
     pub ada_per_utxo_byte: Coin,
-    pub cost_models_for_script_languages: CostMdls,
+    pub cost_models_for_script_languages: CostModels,
     pub execution_costs: ExUnitPrices,
     pub max_tx_ex_units: ExUnits,
     pub max_block_ex_units: ExUnits,
@@ -128,7 +129,7 @@ pub struct BabbageProtParams {
     pub protocol_version: ProtocolVersion,
     pub min_pool_cost: Coin,
     pub ada_per_utxo_byte: Coin,
-    pub cost_models_for_script_languages: BabbageCostMdls,
+    pub cost_models_for_script_languages: BabbageCostModels,
     pub execution_costs: ExUnitPrices,
     pub max_tx_ex_units: ExUnits,
     pub max_block_ex_units: ExUnits,
@@ -156,7 +157,7 @@ pub struct ConwayProtParams {
     pub protocol_version: ProtocolVersion,
     pub min_pool_cost: Coin,
     pub ada_per_utxo_byte: Coin,
-    pub cost_models_for_script_languages: ConwayCostMdls,
+    pub cost_models_for_script_languages: ConwayCostModels,
     pub execution_costs: ExUnitPrices,
     pub max_tx_ex_units: ExUnits,
     pub max_block_ex_units: ExUnits,
