+++ pallas-primitives/src/alonzo/model.rs
@@ -3,18 +3,17 @@
 //! Handcrafted, idiomatic rust artifacts based on based on the [Alonzo CDDL](https://github.com/input-output-hk/cardano-ledger/blob/master/eras/alonzo/test-suite/cddl-files/alonzo.cddl) file in IOHK repo.
 
 use serde::{Deserialize, Serialize};
-use std::{fmt, hash::Hash as StdHash, ops::Deref};
 
-use pallas_codec::minicbor::{data::Tag, Decode, Encode};
-use pallas_crypto::hash::Hash;
+use pallas_codec::minicbor::{self, data::Tag, Decode, Encode};
 
-use pallas_codec::utils::{Bytes, Int, KeepRaw, KeyValuePairs, MaybeIndefArray, Nullable};
-
-// required for derive attrs to work
-use pallas_codec::minicbor;
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub struct VrfCert(#[n(0)] pub Bytes, #[n(1)] pub Bytes);
+pub use crate::{
+    plutus_data::*, AddrKeyhash, AssetName, Bytes, Coin, CostModel, DatumHash, DnsName, Epoch,
+    ExUnitPrices, ExUnits, GenesisDelegateHash, Genesishash, Hash, IPv4, IPv6, Int, KeepRaw,
+    KeyValuePairs, MaybeIndefArray, Metadata, Metadatum, MetadatumLabel, NetworkId, Nonce,
+    NonceVariant, Nullable, PlutusScript, PolicyId, PoolKeyhash, PoolMetadata, PoolMetadataHash,
+    Port, PositiveInterval, ProtocolVersion, RationalNumber, Relay, RewardAccount, ScriptHash,
+    StakeCredential, TransactionIndex, TransactionInput, UnitInterval, VrfCert, VrfKeyhash,
+};
 
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct HeaderBody {
@@ -64,11 +63,6 @@ pub struct HeaderBody {
     pub protocol_minor: u64,
 }
 
-pub type ProtocolVersion = (u64, u64);
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq)]
-pub struct KesSignature {}
-
 pub type MintedHeaderBody<'a> = KeepRaw<'a, HeaderBody>;
 
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
@@ -100,49 +94,9 @@ impl<'a> From<MintedHeaderBody<'a>> for HeaderBody {
     }
 }
 
-#[derive(
-    Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, PartialOrd, Ord, Clone, StdHash,
-)]
-pub struct TransactionInput {
-    #[n(0)]
-    pub transaction_id: Hash<32>,
-
-    #[n(1)]
-    pub index: u64,
-}
-
-// $nonce /= [ 0 // 1, bytes .size 32 ]
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-#[cbor(index_only)]
-pub enum NonceVariant {
-    #[n(0)]
-    NeutralNonce,
-
-    #[n(1)]
-    Nonce,
-}
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub struct Nonce {
-    #[n(0)]
-    pub variant: NonceVariant,
-
-    #[n(1)]
-    pub hash: Option<Hash<32>>,
-}
-
-pub type ScriptHash = Hash<28>;
-
-pub type PolicyId = Hash<28>;
-
-pub type AssetName = Bytes;
-
 pub type Multiasset<A> = KeyValuePairs<PolicyId, KeyValuePairs<AssetName, A>>;
 
 pub type Mint = Multiasset<i64>;
-
-pub type Coin = u64;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
 pub enum Value {
@@ -201,14 +155,8 @@ pub struct TransactionOutput {
     pub amount: Value,
 
     #[n(2)]
-    pub datum_hash: Option<Hash<32>>,
+    pub datum_hash: Option<DatumHash>,
 }
-
-pub type PoolKeyhash = Hash<28>;
-pub type Epoch = u64;
-pub type Genesishash = Bytes;
-pub type GenesisDelegateHash = Bytes;
-pub type VrfKeyhash = Hash<32>;
 
 /* move_instantaneous_reward = [ 0 / 1, { * stake_credential => delta_coin } / coin ]
 ; The first field determines where the funds are drawn from.
@@ -307,179 +255,9 @@ pub struct MoveInstantaneousReward {
     pub target: InstantaneousRewardTarget,
 }
 
-pub type RewardAccount = Bytes;
-
 pub type Withdrawals = KeyValuePairs<RewardAccount, Coin>;
 
 pub type RequiredSigners = Vec<AddrKeyhash>;
-
-pub type Port = u32;
-pub type IPv4 = Bytes;
-pub type IPv6 = Bytes;
-pub type DnsName = String;
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub enum Relay {
-    SingleHostAddr(Nullable<Port>, Nullable<IPv4>, Nullable<IPv6>),
-    SingleHostName(Nullable<Port>, DnsName),
-    MultiHostName(DnsName),
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for Relay {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        d.array()?;
-        let variant = d.u16()?;
-
-        match variant {
-            0 => Ok(Relay::SingleHostAddr(
-                d.decode_with(ctx)?,
-                d.decode_with(ctx)?,
-                d.decode_with(ctx)?,
-            )),
-            1 => Ok(Relay::SingleHostName(
-                d.decode_with(ctx)?,
-                d.decode_with(ctx)?,
-            )),
-            2 => Ok(Relay::MultiHostName(d.decode_with(ctx)?)),
-            _ => Err(minicbor::decode::Error::message(
-                "invalid variant id for Relay",
-            )),
-        }
-    }
-}
-
-impl<C> minicbor::encode::Encode<C> for Relay {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            Relay::SingleHostAddr(a, b, c) => {
-                e.array(4)?;
-                e.encode_with(0, ctx)?;
-                e.encode_with(a, ctx)?;
-                e.encode_with(b, ctx)?;
-                e.encode_with(c, ctx)?;
-
-                Ok(())
-            }
-            Relay::SingleHostName(a, b) => {
-                e.array(3)?;
-                e.encode_with(1, ctx)?;
-                e.encode_with(a, ctx)?;
-                e.encode_with(b, ctx)?;
-
-                Ok(())
-            }
-            Relay::MultiHostName(a) => {
-                e.array(2)?;
-                e.encode_with(2, ctx)?;
-                e.encode_with(a, ctx)?;
-
-                Ok(())
-            }
-        }
-    }
-}
-
-pub type PoolMetadataHash = Hash<32>;
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub struct PoolMetadata {
-    #[n(0)]
-    pub url: String,
-
-    #[n(1)]
-    pub hash: PoolMetadataHash,
-}
-
-pub type AddrKeyhash = Hash<28>;
-pub type Scripthash = Hash<28>;
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
-pub struct RationalNumber {
-    pub numerator: u64,
-    pub denominator: u64,
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for RationalNumber {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        d.tag()?;
-        d.array()?;
-
-        Ok(RationalNumber {
-            numerator: d.decode_with(ctx)?,
-            denominator: d.decode_with(ctx)?,
-        })
-    }
-}
-
-impl<C> minicbor::encode::Encode<C> for RationalNumber {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        // TODO: check if this is the correct tag
-        e.tag(Tag::Unassigned(30))?;
-        e.array(2)?;
-        e.encode_with(self.numerator, ctx)?;
-        e.encode_with(self.denominator, ctx)?;
-
-        Ok(())
-    }
-}
-
-pub type UnitInterval = RationalNumber;
-
-pub type PositiveInterval = RationalNumber;
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, PartialOrd, Eq, Ord, Clone, Hash)]
-pub enum StakeCredential {
-    AddrKeyhash(AddrKeyhash),
-    Scripthash(Scripthash),
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for StakeCredential {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        d.array()?;
-        let variant = d.u16()?;
-
-        match variant {
-            0 => Ok(StakeCredential::AddrKeyhash(d.decode_with(ctx)?)),
-            1 => Ok(StakeCredential::Scripthash(d.decode_with(ctx)?)),
-            _ => Err(minicbor::decode::Error::message(
-                "invalid variant id for StakeCredential",
-            )),
-        }
-    }
-}
-
-impl<C> minicbor::encode::Encode<C> for StakeCredential {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            StakeCredential::AddrKeyhash(a) => {
-                e.array(2)?;
-                e.encode_with(0, ctx)?;
-                e.encode_with(a, ctx)?;
-
-                Ok(())
-            }
-            StakeCredential::Scripthash(a) => {
-                e.array(2)?;
-                e.encode_with(1, ctx)?;
-                e.encode_with(a, ctx)?;
-
-                Ok(())
-            }
-        }
-    }
-}
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone)]
 pub enum Certificate {
@@ -649,17 +427,6 @@ impl<C> minicbor::encode::Encode<C> for Certificate {
     }
 }
 
-#[derive(
-    Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy,
-)]
-#[cbor(index_only)]
-pub enum NetworkId {
-    #[n(0)]
-    One,
-    #[n(1)]
-    Two,
-}
-
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
 #[cbor(index_only)]
 pub enum Language {
@@ -667,9 +434,7 @@ pub enum Language {
     PlutusV1,
 }
 
-pub type CostModel = Vec<i64>;
-
-pub type CostMdls = KeyValuePairs<Language, CostModel>;
+pub type CostModels = KeyValuePairs<Language, CostModel>;
 
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
 #[cbor(map)]
@@ -709,7 +474,7 @@ pub struct ProtocolParamUpdate {
     #[n(17)]
     pub ada_per_utxo_byte: Option<Coin>,
     #[n(18)]
-    pub cost_models_for_script_languages: Option<CostMdls>,
+    pub cost_models_for_script_languages: Option<CostModels>,
     #[n(19)]
     pub execution_costs: Option<ExUnitPrices>,
     #[n(20)]
@@ -862,380 +627,6 @@ impl<C> minicbor::encode::Encode<C> for NativeScript {
     }
 }
 
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-#[cbor(transparent)]
-pub struct PlutusScript(#[n(0)] pub Bytes);
-
-impl AsRef<[u8]> for PlutusScript {
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-}
-
-/// Defined to encode PlutusData bytestring as it is done in the canonical
-/// plutus implementation
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-#[serde(into = "String")]
-#[serde(try_from = "String")]
-pub struct BoundedBytes(Vec<u8>);
-
-impl From<Vec<u8>> for BoundedBytes {
-    fn from(xs: Vec<u8>) -> Self {
-        BoundedBytes(xs)
-    }
-}
-
-impl From<BoundedBytes> for Vec<u8> {
-    fn from(b: BoundedBytes) -> Self {
-        b.0
-    }
-}
-
-impl Deref for BoundedBytes {
-    type Target = Vec<u8>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl TryFrom<String> for BoundedBytes {
-    type Error = hex::FromHexError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        let v = hex::decode(value)?;
-        Ok(BoundedBytes(v))
-    }
-}
-
-impl From<BoundedBytes> for String {
-    fn from(b: BoundedBytes) -> Self {
-        hex::encode(b.deref())
-    }
-}
-
-impl fmt::Display for BoundedBytes {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let bytes: Vec<u8> = self.clone().into();
-
-        f.write_str(&hex::encode(bytes))
-    }
-}
-
-impl<C> Encode<C> for BoundedBytes {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        _: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        // we match the haskell implementation by encoding bytestrings longer than 64
-        // bytes as indefinite lists of bytes
-        const CHUNK_SIZE: usize = 64;
-        let bs: &Vec<u8> = self.deref();
-        if bs.len() <= 64 {
-            e.bytes(bs)?;
-        } else {
-            e.begin_bytes()?;
-            for b in bs.chunks(CHUNK_SIZE) {
-                e.bytes(b)?;
-            }
-            e.end()?;
-        }
-        Ok(())
-    }
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for BoundedBytes {
-    fn decode(d: &mut minicbor::Decoder<'b>, _: &mut C) -> Result<Self, minicbor::decode::Error> {
-        let mut res = Vec::new();
-        for chunk in d.bytes_iter()? {
-            let bs = chunk?;
-            res.extend_from_slice(bs);
-        }
-        Ok(BoundedBytes::from(res))
-    }
-}
-
-/*
-big_int = int / big_uint / big_nint ; New
-big_uint = #6.2(bounded_bytes) ; New
-big_nint = #6.3(bounded_bytes) ; New
- */
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
-pub enum BigInt {
-    Int(Int),
-    BigUInt(BoundedBytes),
-    BigNInt(BoundedBytes),
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for BigInt {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        let datatype = d.datatype()?;
-
-        match datatype {
-            minicbor::data::Type::U8
-            | minicbor::data::Type::U16
-            | minicbor::data::Type::U32
-            | minicbor::data::Type::U64
-            | minicbor::data::Type::I8
-            | minicbor::data::Type::I16
-            | minicbor::data::Type::I32
-            | minicbor::data::Type::I64
-            | minicbor::data::Type::Int => Ok(Self::Int(d.decode_with(ctx)?)),
-            minicbor::data::Type::Tag => {
-                let tag = d.tag()?;
-
-                match tag {
-                    minicbor::data::Tag::PosBignum => Ok(Self::BigUInt(d.decode_with(ctx)?)),
-                    minicbor::data::Tag::NegBignum => Ok(Self::BigNInt(d.decode_with(ctx)?)),
-                    _ => Err(minicbor::decode::Error::message(
-                        "invalid cbor tag for big int",
-                    )),
-                }
-            }
-            _ => Err(minicbor::decode::Error::message(
-                "invalid cbor data type for big int",
-            )),
-        }
-    }
-}
-
-impl<C> minicbor::encode::Encode<C> for BigInt {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            BigInt::Int(x) => {
-                e.encode_with(x, ctx)?;
-            }
-            BigInt::BigUInt(x) => {
-                e.tag(Tag::PosBignum)?;
-                e.encode_with(x, ctx)?;
-            }
-            BigInt::BigNInt(x) => {
-                e.tag(Tag::NegBignum)?;
-                e.encode_with(x, ctx)?;
-            }
-        };
-
-        Ok(())
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
-pub enum PlutusData {
-    Constr(Constr<PlutusData>),
-    Map(KeyValuePairs<PlutusData, PlutusData>),
-    BigInt(BigInt),
-    BoundedBytes(BoundedBytes),
-    Array(Vec<PlutusData>),
-}
-
-impl<'b, C> minicbor::decode::Decode<'b, C> for PlutusData {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        let type_ = d.datatype()?;
-
-        match type_ {
-            minicbor::data::Type::Tag => {
-                let mut probe = d.probe();
-                let tag = probe.tag()?;
-
-                match tag {
-                    Tag::Unassigned(121..=127 | 1280..=1400 | 102) => {
-                        Ok(Self::Constr(d.decode_with(ctx)?))
-                    }
-                    Tag::PosBignum | Tag::NegBignum => Ok(Self::BigInt(d.decode_with(ctx)?)),
-                    _ => Err(minicbor::decode::Error::message(
-                        "unknown tag for plutus data tag",
-                    )),
-                }
-            }
-            minicbor::data::Type::U8
-            | minicbor::data::Type::U16
-            | minicbor::data::Type::U32
-            | minicbor::data::Type::U64
-            | minicbor::data::Type::I8
-            | minicbor::data::Type::I16
-            | minicbor::data::Type::I32
-            | minicbor::data::Type::I64
-            | minicbor::data::Type::Int => Ok(Self::BigInt(d.decode_with(ctx)?)),
-            minicbor::data::Type::Map | minicbor::data::Type::MapIndef => {
-                Ok(Self::Map(d.decode_with(ctx)?))
-            }
-            minicbor::data::Type::Bytes => Ok(Self::BoundedBytes(d.decode_with(ctx)?)),
-            minicbor::data::Type::BytesIndef => {
-                let mut full = Vec::new();
-
-                for slice in d.bytes_iter()? {
-                    full.extend(slice?);
-                }
-
-                Ok(Self::BoundedBytes(BoundedBytes::from(full)))
-            }
-            minicbor::data::Type::Array | minicbor::data::Type::ArrayIndef => {
-                Ok(Self::Array(d.decode_with(ctx)?))
-            }
-
-            any => Err(minicbor::decode::Error::message(format!(
-                "bad cbor data type ({any:?}) for plutus data"
-            ))),
-        }
-    }
-}
-
-fn encode_list<C, W: minicbor::encode::Write, A: minicbor::encode::Encode<C>>(
-    a: &Vec<A>,
-    e: &mut minicbor::Encoder<W>,
-    ctx: &mut C,
-) -> Result<(), minicbor::encode::Error<W::Error>> {
-    // Mimics default haskell list encoding from cborg:
-    // We use indef array for non-empty arrays but definite 0-length array for empty
-    if a.is_empty() {
-        e.array(0)?;
-    } else {
-        e.begin_array()?;
-        for v in a {
-            e.encode_with(v, ctx)?;
-        }
-        e.end()?;
-    }
-    Ok(())
-}
-
-impl<C> minicbor::encode::Encode<C> for PlutusData {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            Self::Constr(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Self::Map(a) => {
-                // we use definite array to match the approach used by haskell's plutus
-                // implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L152
-                e.map(a.len().try_into().unwrap())?;
-                for (k, v) in a.iter() {
-                    k.encode(e, ctx)?;
-                    v.encode(e, ctx)?;
-                }
-            }
-            Self::BigInt(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Self::BoundedBytes(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Self::Array(a) => {
-                // we use definite array for empty array or indef array otherwise to match
-                // haskell implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L153
-                // default encoder for a list:
-                // https://github.com/well-typed/cborg/blob/4bdc818a1f0b35f38bc118a87944630043b58384/serialise/src/Codec/Serialise/Class.hs#L181
-                encode_list(a, e, ctx)?;
-            }
-        };
-
-        Ok(())
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
-pub struct Constr<A> {
-    pub tag: u64,
-    pub any_constructor: Option<u64>,
-    pub fields: Vec<A>,
-}
-
-impl<'b, C, A> minicbor::decode::Decode<'b, C> for Constr<A>
-where
-    A: minicbor::decode::Decode<'b, C>,
-{
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        let tag = d.tag()?;
-
-        match tag {
-            Tag::Unassigned(x) => match x {
-                121..=127 | 1280..=1400 => Ok(Constr {
-                    tag: x,
-                    fields: d.decode_with(ctx)?,
-                    any_constructor: None,
-                }),
-                102 => {
-                    d.array()?;
-
-                    Ok(Constr {
-                        tag: x,
-                        any_constructor: Some(d.decode_with(ctx)?),
-                        fields: d.decode_with(ctx)?,
-                    })
-                }
-                _ => Err(minicbor::decode::Error::message(
-                    "bad tag code for plutus data",
-                )),
-            },
-            _ => Err(minicbor::decode::Error::message(
-                "bad tag code for plutus data",
-            )),
-        }
-    }
-}
-
-impl<C, A> minicbor::encode::Encode<C> for Constr<A>
-where
-    A: minicbor::encode::Encode<C>,
-{
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        e.tag(Tag::Unassigned(self.tag))?;
-
-        match self.tag {
-            102 => {
-                // definite length array here
-                // https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L152
-                e.array(2)?;
-                e.encode_with(self.any_constructor.unwrap_or_default(), ctx)?;
-
-                // we use definite array for empty array or indef array otherwise to match
-                // haskell implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L144
-                // default encoder for a list:
-                // https://github.com/well-typed/cborg/blob/4bdc818a1f0b35f38bc118a87944630043b58384/serialise/src/Codec/Serialise/Class.hs#L181
-                encode_list(&self.fields, e, ctx)?;
-                Ok(())
-            }
-            _ => {
-                // we use definite array for empty array or indef array otherwise to match
-                // haskell implementation. See above reference.
-                encode_list(&self.fields, e, ctx)?;
-                Ok(())
-            }
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone, Copy)]
-pub struct ExUnits {
-    #[n(0)]
-    pub mem: u64,
-    #[n(1)]
-    pub steps: u64,
-}
-
-#[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub struct ExUnitPrices {
-    #[n(0)]
-    pub mem_price: PositiveInterval,
-
-    #[n(1)]
-    pub step_price: PositiveInterval,
-}
-
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Eq, Clone, Copy)]
 #[cbor(index_only)]
 pub enum RedeemerTag {
@@ -1308,7 +699,7 @@ pub struct WitnessSet {
     pub bootstrap_witness: Option<Vec<BootstrapWitness>>,
 
     #[n(3)]
-    pub plutus_script: Option<Vec<PlutusScript>>,
+    pub plutus_v1_script: Option<Vec<PlutusScript<1>>>,
 
     #[n(4)]
     pub plutus_data: Option<Vec<PlutusData>>,
@@ -1330,7 +721,7 @@ pub struct MintedWitnessSet<'b> {
     pub bootstrap_witness: Option<Vec<BootstrapWitness>>,
 
     #[n(3)]
-    pub plutus_script: Option<Vec<PlutusScript>>,
+    pub plutus_v1_script: Option<Vec<PlutusScript<1>>>,
 
     #[b(4)]
     pub plutus_data: Option<Vec<KeepRaw<'b, PlutusData>>>,
@@ -1347,7 +738,7 @@ impl<'b> From<MintedWitnessSet<'b>> for WitnessSet {
                 .native_script
                 .map(|x| x.into_iter().map(|x| x.unwrap()).collect()),
             bootstrap_witness: x.bootstrap_witness,
-            plutus_script: x.plutus_script,
+            plutus_v1_script: x.plutus_v1_script,
             plutus_data: x
                 .plutus_data
                 .map(|x| x.into_iter().map(|x| x.unwrap()).collect()),
@@ -1366,79 +757,8 @@ pub struct PostAlonzoAuxiliaryData {
     pub native_scripts: Option<Vec<NativeScript>>,
 
     #[n(2)]
-    pub plutus_scripts: Option<Vec<PlutusScript>>,
+    pub plutus_scripts: Option<Vec<PlutusScript<1>>>,
 }
-
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
-pub enum Metadatum {
-    Int(Int),
-    Bytes(Bytes),
-    Text(String),
-    Array(Vec<Metadatum>),
-    Map(KeyValuePairs<Metadatum, Metadatum>),
-}
-
-impl<'b, C> minicbor::Decode<'b, C> for Metadatum {
-    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
-        match d.datatype()? {
-            minicbor::data::Type::U8
-            | minicbor::data::Type::U16
-            | minicbor::data::Type::U32
-            | minicbor::data::Type::U64
-            | minicbor::data::Type::I8
-            | minicbor::data::Type::I16
-            | minicbor::data::Type::I32
-            | minicbor::data::Type::I64
-            | minicbor::data::Type::Int => {
-                let i = d.decode()?;
-                Ok(Metadatum::Int(i))
-            }
-            minicbor::data::Type::Bytes => Ok(Metadatum::Bytes(d.decode_with(ctx)?)),
-            minicbor::data::Type::String => Ok(Metadatum::Text(d.decode_with(ctx)?)),
-            minicbor::data::Type::Array | minicbor::data::Type::ArrayIndef => {
-                Ok(Metadatum::Array(d.decode_with(ctx)?))
-            }
-            minicbor::data::Type::Map | minicbor::data::Type::MapIndef => {
-                Ok(Metadatum::Map(d.decode_with(ctx)?))
-            }
-            _ => Err(minicbor::decode::Error::message(
-                "Can't turn data type into metadatum",
-            )),
-        }
-    }
-}
-
-impl<C> minicbor::Encode<C> for Metadatum {
-    fn encode<W: minicbor::encode::Write>(
-        &self,
-        e: &mut minicbor::Encoder<W>,
-        ctx: &mut C,
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            Metadatum::Int(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Metadatum::Bytes(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Metadatum::Text(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Metadatum::Array(a) => {
-                e.encode_with(a, ctx)?;
-            }
-            Metadatum::Map(a) => {
-                e.encode_with(a, ctx)?;
-            }
-        };
-
-        Ok(())
-    }
-}
-
-pub type MetadatumLabel = u64;
-
-pub type Metadata = KeyValuePairs<MetadatumLabel, Metadatum>;
 
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Clone)]
 pub struct ShelleyMaAuxiliaryData {
@@ -1497,8 +817,6 @@ impl<C> minicbor::Encode<C> for AuxiliaryData {
         Ok(())
     }
 }
-
-pub type TransactionIndex = u32;
 
 #[derive(Serialize, Deserialize, Encode, Decode, Debug, PartialEq, Clone)]
 pub struct Block {
