+++ pallas-primitives/src/plutus_data.rs
@@ -1,0 +1,352 @@
+use crate::KeyValuePairs;
+use pallas_codec::minicbor::{self, data::Tag, Encode};
+use pallas_codec::utils::Int;
+use serde::{Deserialize, Serialize};
+use std::{fmt, ops::Deref};
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub enum PlutusData {
+    Constr(Constr<PlutusData>),
+    Map(KeyValuePairs<PlutusData, PlutusData>),
+    BigInt(BigInt),
+    BoundedBytes(BoundedBytes),
+    Array(Vec<PlutusData>),
+}
+
+impl<'b, C> minicbor::decode::Decode<'b, C> for PlutusData {
+    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
+        let type_ = d.datatype()?;
+
+        match type_ {
+            minicbor::data::Type::Tag => {
+                let mut probe = d.probe();
+                let tag = probe.tag()?;
+
+                match tag {
+                    Tag::Unassigned(121..=127 | 1280..=1400 | 102) => {
+                        Ok(Self::Constr(d.decode_with(ctx)?))
+                    }
+                    Tag::PosBignum | Tag::NegBignum => Ok(Self::BigInt(d.decode_with(ctx)?)),
+                    _ => Err(minicbor::decode::Error::message(
+                        "unknown tag for plutus data tag",
+                    )),
+                }
+            }
+            minicbor::data::Type::U8
+            | minicbor::data::Type::U16
+            | minicbor::data::Type::U32
+            | minicbor::data::Type::U64
+            | minicbor::data::Type::I8
+            | minicbor::data::Type::I16
+            | minicbor::data::Type::I32
+            | minicbor::data::Type::I64
+            | minicbor::data::Type::Int => Ok(Self::BigInt(d.decode_with(ctx)?)),
+            minicbor::data::Type::Map | minicbor::data::Type::MapIndef => {
+                Ok(Self::Map(d.decode_with(ctx)?))
+            }
+            minicbor::data::Type::Bytes => Ok(Self::BoundedBytes(d.decode_with(ctx)?)),
+            minicbor::data::Type::BytesIndef => {
+                let mut full = Vec::new();
+
+                for slice in d.bytes_iter()? {
+                    full.extend(slice?);
+                }
+
+                Ok(Self::BoundedBytes(BoundedBytes::from(full)))
+            }
+            minicbor::data::Type::Array | minicbor::data::Type::ArrayIndef => {
+                Ok(Self::Array(d.decode_with(ctx)?))
+            }
+
+            any => Err(minicbor::decode::Error::message(format!(
+                "bad cbor data type ({any:?}) for plutus data"
+            ))),
+        }
+    }
+}
+
+impl<C> minicbor::encode::Encode<C> for PlutusData {
+    fn encode<W: minicbor::encode::Write>(
+        &self,
+        e: &mut minicbor::Encoder<W>,
+        ctx: &mut C,
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        match self {
+            Self::Constr(a) => {
+                e.encode_with(a, ctx)?;
+            }
+            Self::Map(a) => {
+                // we use definite array to match the approach used by haskell's plutus
+                // implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L152
+                e.map(a.len().try_into().unwrap())?;
+                for (k, v) in a.iter() {
+                    k.encode(e, ctx)?;
+                    v.encode(e, ctx)?;
+                }
+            }
+            Self::BigInt(a) => {
+                e.encode_with(a, ctx)?;
+            }
+            Self::BoundedBytes(a) => {
+                e.encode_with(a, ctx)?;
+            }
+            Self::Array(a) => {
+                // we use definite array for empty array or indef array otherwise to match
+                // haskell implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L153
+                // default encoder for a list:
+                // https://github.com/well-typed/cborg/blob/4bdc818a1f0b35f38bc118a87944630043b58384/serialise/src/Codec/Serialise/Class.hs#L181
+                encode_list(a, e, ctx)?;
+            }
+        };
+
+        Ok(())
+    }
+}
+
+/*
+big_int = int / big_uint / big_nint ; New
+big_uint = #6.2(bounded_bytes) ; New
+big_nint = #6.3(bounded_bytes) ; New
+ */
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub enum BigInt {
+    Int(Int),
+    BigUInt(BoundedBytes),
+    BigNInt(BoundedBytes),
+}
+
+impl<'b, C> minicbor::decode::Decode<'b, C> for BigInt {
+    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
+        let datatype = d.datatype()?;
+
+        match datatype {
+            minicbor::data::Type::U8
+            | minicbor::data::Type::U16
+            | minicbor::data::Type::U32
+            | minicbor::data::Type::U64
+            | minicbor::data::Type::I8
+            | minicbor::data::Type::I16
+            | minicbor::data::Type::I32
+            | minicbor::data::Type::I64
+            | minicbor::data::Type::Int => Ok(Self::Int(d.decode_with(ctx)?)),
+            minicbor::data::Type::Tag => {
+                let tag = d.tag()?;
+
+                match tag {
+                    minicbor::data::Tag::PosBignum => Ok(Self::BigUInt(d.decode_with(ctx)?)),
+                    minicbor::data::Tag::NegBignum => Ok(Self::BigNInt(d.decode_with(ctx)?)),
+                    _ => Err(minicbor::decode::Error::message(
+                        "invalid cbor tag for big int",
+                    )),
+                }
+            }
+            _ => Err(minicbor::decode::Error::message(
+                "invalid cbor data type for big int",
+            )),
+        }
+    }
+}
+
+impl<C> minicbor::encode::Encode<C> for BigInt {
+    fn encode<W: minicbor::encode::Write>(
+        &self,
+        e: &mut minicbor::Encoder<W>,
+        ctx: &mut C,
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        match self {
+            BigInt::Int(x) => {
+                e.encode_with(x, ctx)?;
+            }
+            BigInt::BigUInt(x) => {
+                e.tag(Tag::PosBignum)?;
+                e.encode_with(x, ctx)?;
+            }
+            BigInt::BigNInt(x) => {
+                e.tag(Tag::NegBignum)?;
+                e.encode_with(x, ctx)?;
+            }
+        };
+
+        Ok(())
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub struct Constr<A> {
+    pub tag: u64,
+    pub any_constructor: Option<u64>,
+    pub fields: Vec<A>,
+}
+
+impl<'b, C, A> minicbor::decode::Decode<'b, C> for Constr<A>
+where
+    A: minicbor::decode::Decode<'b, C>,
+{
+    fn decode(d: &mut minicbor::Decoder<'b>, ctx: &mut C) -> Result<Self, minicbor::decode::Error> {
+        let tag = d.tag()?;
+
+        match tag {
+            Tag::Unassigned(x) => match x {
+                121..=127 | 1280..=1400 => Ok(Constr {
+                    tag: x,
+                    fields: d.decode_with(ctx)?,
+                    any_constructor: None,
+                }),
+                102 => {
+                    d.array()?;
+
+                    Ok(Constr {
+                        tag: x,
+                        any_constructor: Some(d.decode_with(ctx)?),
+                        fields: d.decode_with(ctx)?,
+                    })
+                }
+                _ => Err(minicbor::decode::Error::message(
+                    "bad tag code for plutus data",
+                )),
+            },
+            _ => Err(minicbor::decode::Error::message(
+                "bad tag code for plutus data",
+            )),
+        }
+    }
+}
+
+impl<C, A> minicbor::encode::Encode<C> for Constr<A>
+where
+    A: minicbor::encode::Encode<C>,
+{
+    fn encode<W: minicbor::encode::Write>(
+        &self,
+        e: &mut minicbor::Encoder<W>,
+        ctx: &mut C,
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        e.tag(Tag::Unassigned(self.tag))?;
+
+        match self.tag {
+            102 => {
+                // definite length array here
+                // https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L152
+                e.array(2)?;
+                e.encode_with(self.any_constructor.unwrap_or_default(), ctx)?;
+
+                // we use definite array for empty array or indef array otherwise to match
+                // haskell implementation https://github.com/input-output-hk/plutus/blob/9538fc9829426b2ecb0628d352e2d7af96ec8204/plutus-core/plutus-core/src/PlutusCore/Data.hs#L144
+                // default encoder for a list:
+                // https://github.com/well-typed/cborg/blob/4bdc818a1f0b35f38bc118a87944630043b58384/serialise/src/Codec/Serialise/Class.hs#L181
+                encode_list(&self.fields, e, ctx)?;
+                Ok(())
+            }
+            _ => {
+                // we use definite array for empty array or indef array otherwise to match
+                // haskell implementation. See above reference.
+                encode_list(&self.fields, e, ctx)?;
+                Ok(())
+            }
+        }
+    }
+}
+
+/// Defined to encode PlutusData bytestring as it is done in the canonical
+/// plutus implementation
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+#[serde(into = "String")]
+#[serde(try_from = "String")]
+pub struct BoundedBytes(Vec<u8>);
+
+impl From<Vec<u8>> for BoundedBytes {
+    fn from(xs: Vec<u8>) -> Self {
+        BoundedBytes(xs)
+    }
+}
+
+impl From<BoundedBytes> for Vec<u8> {
+    fn from(b: BoundedBytes) -> Self {
+        b.0
+    }
+}
+
+impl Deref for BoundedBytes {
+    type Target = Vec<u8>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl TryFrom<String> for BoundedBytes {
+    type Error = hex::FromHexError;
+
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        let v = hex::decode(value)?;
+        Ok(BoundedBytes(v))
+    }
+}
+
+impl From<BoundedBytes> for String {
+    fn from(b: BoundedBytes) -> Self {
+        hex::encode(b.deref())
+    }
+}
+
+impl fmt::Display for BoundedBytes {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let bytes: Vec<u8> = self.clone().into();
+
+        f.write_str(&hex::encode(bytes))
+    }
+}
+
+impl<C> Encode<C> for BoundedBytes {
+    fn encode<W: minicbor::encode::Write>(
+        &self,
+        e: &mut minicbor::Encoder<W>,
+        _: &mut C,
+    ) -> Result<(), minicbor::encode::Error<W::Error>> {
+        // we match the haskell implementation by encoding bytestrings longer than 64
+        // bytes as indefinite lists of bytes
+        const CHUNK_SIZE: usize = 64;
+        let bs: &Vec<u8> = self.deref();
+        if bs.len() <= 64 {
+            e.bytes(bs)?;
+        } else {
+            e.begin_bytes()?;
+            for b in bs.chunks(CHUNK_SIZE) {
+                e.bytes(b)?;
+            }
+            e.end()?;
+        }
+        Ok(())
+    }
+}
+
+impl<'b, C> minicbor::decode::Decode<'b, C> for BoundedBytes {
+    fn decode(d: &mut minicbor::Decoder<'b>, _: &mut C) -> Result<Self, minicbor::decode::Error> {
+        let mut res = Vec::new();
+        for chunk in d.bytes_iter()? {
+            let bs = chunk?;
+            res.extend_from_slice(bs);
+        }
+        Ok(BoundedBytes::from(res))
+    }
+}
+
+fn encode_list<C, W: minicbor::encode::Write, A: minicbor::encode::Encode<C>>(
+    a: &Vec<A>,
+    e: &mut minicbor::Encoder<W>,
+    ctx: &mut C,
+) -> Result<(), minicbor::encode::Error<W::Error>> {
+    // Mimics default haskell list encoding from cborg:
+    // We use indef array for non-empty arrays but definite 0-length array for empty
+    if a.is_empty() {
+        e.array(0)?;
+    } else {
+        e.begin_array()?;
+        for v in a {
+            e.encode_with(v, ctx)?;
+        }
+        e.end()?;
+    }
+    Ok(())
+}
