+++ pallas-traverse/src/hashes.rs
@@ -82,9 +82,9 @@ impl OriginalHash<28> for KeepRaw<'_, alonzo::NativeScript> {
     }
 }
 
-impl ComputeHash<28> for alonzo::PlutusScript {
+impl<const VERSION: usize> ComputeHash<28> for alonzo::PlutusScript<VERSION> {
     fn compute_hash(&self) -> Hash<28> {
-        Hasher::<224>::hash_tagged(&self.0, 1)
+        Hasher::<224>::hash_tagged(&self.0, VERSION as u8)
     }
 }
 
@@ -124,12 +124,6 @@ impl OriginalHash<32> for KeepRaw<'_, babbage::Header> {
     }
 }
 
-impl ComputeHash<28> for babbage::PlutusV2Script {
-    fn compute_hash(&self) -> Hash<28> {
-        Hasher::<224>::hash_tagged(&self.0, 2)
-    }
-}
-
 impl ComputeHash<32> for babbage::TransactionBody {
     fn compute_hash(&self) -> Hash<32> {
         Hasher::<256>::hash_cbor(self)
@@ -158,12 +152,6 @@ impl ComputeHash<32> for babbage::DatumOption {
 }
 
 // conway
-
-impl ComputeHash<28> for conway::PlutusV3Script {
-    fn compute_hash(&self) -> Hash<28> {
-        Hasher::<224>::hash_tagged(&self.0, 3)
-    }
-}
 
 impl ComputeHash<32> for conway::TransactionBody {
     fn compute_hash(&self) -> Hash<32> {
@@ -334,7 +322,7 @@ mod tests {
     fn plutus_v1_script_hashes_as_cardano_cli() {
         let bytecode_hex = include_str!("../../test_data/jpgstore.plutus");
         let bytecode = hex::decode(bytecode_hex).unwrap();
-        let script = alonzo::PlutusScript(Bytes::from(bytecode));
+        let script = alonzo::PlutusScript::<1>(Bytes::from(bytecode));
 
         let generated = script.compute_hash().to_string();
 
@@ -350,7 +338,7 @@ mod tests {
     fn plutus_v2_script_hashes_as_cardano_cli() {
         let bytecode_hex = include_str!("../../test_data/v2script.plutus");
         let bytecode = hex::decode(bytecode_hex).unwrap();
-        let script = babbage::PlutusV2Script(Bytes::from(bytecode));
+        let script = babbage::PlutusScript::<2>(Bytes::from(bytecode));
 
         let generated = script.compute_hash().to_string();
